# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kinoage",
        "description": "Documentation for kinoage admin dashboard and main site",
        "contact": {
            "email": "ayubxonnematov@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of User",
                "description": "Returns list of User",
                "operationId": "indexUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Insert a new User",
                "description": "Insert a new User",
                "operationId": "storeUser",
                "requestBody": {
                    "description": "User to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/user/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show a User from his Id",
                "description": "Show a User from his Id",
                "operationId": "showUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/User--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a User",
                "description": "Delete a User",
                "operationId": "destroyUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/User--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update an existing User",
                "description": "Update an existing User",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/User--id"
                    }
                ],
                "requestBody": {
                    "description": "User to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/role": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Show a role list",
                "description": "Show a role list",
                "operationId": "showRole",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of User",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "first_name": {
                        "title": "first_name",
                        "description": "first_name of User",
                        "type": "string",
                        "example": "Lorem"
                    },
                    "last_name": {
                        "title": "last_name",
                        "description": "last_name of User",
                        "type": "string",
                        "example": "Lorem"
                    },
                    "middle_name": {
                        "title": "middle_name",
                        "description": "middle_name of User",
                        "type": "string",
                        "example": "Lorem"
                    },
                    "username": {
                        "title": "username",
                        "description": "username of User",
                        "type": "string",
                        "example": "example"
                    },
                    "roles": {
                        "title": "roles",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "title": "id",
                                    "description": "id of Role",
                                    "type": "integer",
                                    "readOnly": "true",
                                    "example": "1"
                                },
                                "role": {
                                    "title": "role",
                                    "description": "Name of Role",
                                    "type": "string",
                                    "readOnly": "true",
                                    "example": "Merchant"
                                },
                                "status": {
                                    "title": "status",
                                    "description": "Status of Role",
                                    "type": "integer",
                                    "readOnly": "true",
                                    "example": "20"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
            "Users": {
                "title": "Users",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    },
                    "links": {
                        "title": "links",
                        "properties": {
                            "first": {
                                "title": "first",
                                "type": "string",
                                "example": "http://localhost:8080/api/users?page=1"
                            },
                            "last": {
                                "title": "last",
                                "type": "string",
                                "example": "http://localhost:8080/api/users?page=3"
                            },
                            "prev": {
                                "title": "prev",
                                "type": "string",
                                "example": "null"
                            },
                            "next": {
                                "title": "next",
                                "type": "string",
                                "example": "http://localhost:8080/api/users?page=2"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "title": "meta",
                        "properties": {
                            "current_page": {
                                "title": "current_page",
                                "type": "integer",
                                "example": "1"
                            },
                            "from": {
                                "title": "from",
                                "type": "integer",
                                "example": "1"
                            },
                            "last_page": {
                                "title": "last_page",
                                "type": "integer",
                                "example": "3"
                            },
                            "path": {
                                "title": "path",
                                "type": "string",
                                "example": "http://localhost:8080/api/users"
                            },
                            "per_page": {
                                "title": "per_page",
                                "type": "integer",
                                "example": "1"
                            },
                            "total": {
                                "title": "total",
                                "type": "integer",
                                "example": "3"
                            },
                            "links": {
                                "title": "links",
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "url": {
                                            "title": "url",
                                            "type": "string",
                                            "example": "http://localhost:8080/api/users?page=2"
                                        },
                                        "label": {
                                            "title": "label",
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "active": {
                                            "title": "active",
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "User--id": {
                "name": "user_id",
                "in": "path",
                "description": "Id of User",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}